% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_api.R
\name{sim_pacbio}
\alias{sim_pacbio}
\title{Simulate PacBio reads with PBSIM/PBSIM3}
\usage{
sim_pacbio(ref_fa, outdir, cov, type = "HIFI", seed = 1)
}
\arguments{
\item{ref_fa}{Character. Reference FASTA path.}

\item{outdir}{Character. Output directory.}

\item{cov}{Numeric. Fold coverage.}

\item{type}{Character. \code{"CLR"} or \code{"HIFI"}.}

\item{seed}{Integer. Random seed (default 1).}
}
\value{
Character. Output directory path.
}
\description{
Wrapper around \code{pbsim} or \code{pbsim3} for CLR/HiFi simulation.
}
\examples{
\dontrun{
sim_pacbio("01_simref/ecoli_repMed.fa", "02_reads/ecoli/pacbio_HIFI/cov20", 20, type = "HIFI")
}
}
