% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitall-package.R
\docType{package}
\name{simitall-package}
\alias{simitall}
\alias{simitall-package}
\title{simitall: Integrated genomics simulation from R}
\description{
\code{simitall} coordinates genome, annotation, sequencing, assembly, GWAS,
bulk and single-cell RNA-seq, ChIP-seq, eQTL, phenotype, population, and
breeding simulations while retaining explicit truth outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nirwan1265/simitall}
  \item Report bugs at \url{https://github.com/nirwan1265/simitall/issues}
}

}
\author{
\strong{Maintainer}: Nirwan Tandukar \email{tandukar.nirwan@gmail.com}

}
\keyword{internal}
